backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.115.0
sqlmodel>=0.0.22
pydantic>=2.10.0
pydantic-settings>=2.7.0
uvicorn[standard]>=0.34.0

# Database (psycopg v3 - modern async-capable PostgreSQL driver)
psycopg[binary]>=3.1

# Environment configuration
python-dotenv>=1.0.0

# Validation
email-validator>=2.0.0

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Development/Testing (optional)
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
alembic>=1.13.1
